requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.22
sqlalchemy-utils==0.41.1
python-multipart==0.0.6
httpx==0.24.1
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
requests==2.31.0
pytest==7.4.3
openai>=1.25.0
pypdf==5.4.0 